Draft blob module, blob test, minor refactoring on the util.py and hashing.py, initial placeholders for core modules

+++ common/util.py
@@ -7,6 +7,7 @@ from pprint import pprint
 from . import config
 
 CFG_DICT = config.CFG_DICT
+LOG_PREFIX = CFG_DICT['LOG_PREFIX']
 
 
 def _time_measure(func):
@@ -21,18 +22,35 @@ def _time_measure(func):
     return wrapper
 
 
+def log_info(*args):
+    _log_indent = 0
+    message = ''.join(args)
+    indent_str = ''.join([' ' for i in range(_log_indent * 4)])
+    print('{} - {}'.format(LOG_PREFIX, indent_str + message))
+
+
 def match_regex_pattern(input_str, patterns):
     for p in patterns:
         if re.findall(p, input_str):
             return 1
 
 
-def copy_file(source, target):
+def copy_file(source, target, verbose=0):
     if not Path(target).parent.exists():
         Path(target).parent.mkdir(parents=1)
-    shutil.copyfile(str(source), str(target))
+    if not Path(target).exists():
+        shutil.copyfile(str(source), str(target))
+        if verbose:
+            log_info('Copied: {}'.format(target))
 
 
-def batch_copy(path_pair):
+def batch_copy(path_pair, verbose=0):
+    """
+    Params
+    ------
+    path_pair : list
+        List of 2-tuple of str, [(source, target),...]
+    """
+
     for source, target in path_pair:
-        copy_file(source, target)
+        copy_file(source, target, verbose=verbose)
